% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccor_boot.R
\name{ccor_boot}
\alias{ccor_boot}
\title{Bootstrap confidence interval estimation for curve correlation}
\usage{
ccor_boot(
  y,
  time,
  curve = NULL,
  k = 15,
  ndraw = 299,
  conf = 0.95,
  min.overlap = 0,
  min.n = 8
)
}
\arguments{
\item{y, time, curve, k, min.overlap, min.n}{see \code{\link{ccor}}}

\item{ndraw}{number of bootstrap samples}

\item{conf}{confidence level}
}
\value{
A list with components
\item{cor }{curve correlation (for the full data)}
 \item{boot.cor }{curve correlations for the \code{ndraw} bootstrap samples}
 \item{ci }{confidence interval for the curve correlation}
}
\description{
Inputs raw data representing two curves, and computes a bootstrap confidence interval
for the curve correlation between them.
}
\examples{

# see example for ccor_posim

}
\seealso{
\code{\link{ccor}}, \code{\link{ccor_posim}}
}
\author{
Philip Tzvi Reiss <reiss@stat.haifa.ac.il>, Noemi Foa, Dror Arbiv and Biplab Paul <paul.biplab497@gmail.com>
}
