% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccor_oneboot.R
\name{ccor_oneboot}
\alias{ccor_oneboot}
\title{Curve correlation for a single bootstrap sample}
\usage{
ccor_oneboot(y, time, curve = NULL, k = 15, min.overlap = 0, min.n = 8)
}
\arguments{
\item{y, time, curve, k, min.overlap, min.n}{see \code{\link{ccor}}}
}
\description{
This internal function is called repeatedly by \code{\link{ccor_boot}}. It is not ordinarily called by users.
}
\author{
Philip Tzvi Reiss <reiss@stat.haifa.ac.il>, Noemi Foa, Dror Arbiv and Biplab Paul <paul.biplab497@gmail.com>
}
\keyword{internal}
