% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.ct.R
\name{center.ct}
\alias{center.ct}
\title{Center a continuous-time multivariate data set}
\usage{
center.ct(fdobj)
}
\arguments{
\item{fdobj}{continuous-time multivariate data set of class \code{"\link[fda]{fd}"}}
}
\value{
A centered version of the input data.
}
\description{
Subtracts the (continuous-time) mean of each of the variables. This is analogous
to column-centering an \eqn{n \times p} data matrix.
}
\seealso{
\code{\link{standardize.ct}}
}
\author{
Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
