% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.ct.R
\name{cov.ct}
\alias{cov.ct}
\title{Continuous-time covariance or cross-covariance matrix}
\usage{
cov.ct(fdobj1, fdobj2 = fdobj1, common_trend = FALSE)
}
\arguments{
\item{fdobj1}{continuous-time multivariate data set of class \code{"\link[fda]{fd}"}}

\item{fdobj2}{an optional second data set}

\item{common_trend}{logical: centering with respect to the mean function if \code{TRUE},
without centering if \code{FALSE} (the default)}
}
\value{
A matrix of (cross-) covariances
}
\description{
Computes the covariance matrix of a continuous-time multivariate
data set represented as an \code{\link[fda]{fd}} object; or the
cross-covariance matrix of two such data sets.
}
\examples{

# see example for cor.ct, which works similarly

}
\seealso{
\code{\link{cor.ct}}
}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
