% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctmva-internal}
\alias{ctmva-internal}
\alias{mygen}
\alias{trapwts}
\title{Internal helper functions for ctmva}
\usage{
mygen(mtrx, Im.tol = 1e-15)

trapwts(tt)
}
\description{
Internal helper functions for ctmva
}
\section{Functions}{
\itemize{
\item \code{mygen()}: Modified version of eigen

\item \code{trapwts()}: Trapezoid weights

}}
\keyword{internal}
