% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lda.ct.R
\name{plot.lda.ct}
\alias{plot.lda.ct}
\title{Plot an lda.ct object}
\usage{
\method{plot}{lda.ct}(x, ylab = "Discriminants", xlab = "Time", ...)
}
\arguments{
\item{x}{linear discriminant analysis object produced by \code{\link{lda.ct}}}

\item{ylab, xlab}{y- and x-axis labels}

\item{\dots}{other arguments passed to \code{\link{matplot}}}
}
\value{
None; a plot is generated.
}
\description{
Plots the Fisher's linear discriminant functions generated by a call
to \code{\link{lda.ct}}.
}
\examples{

## see the example at the end of ?pca.ct


}
\seealso{
\code{\link{lda.ct}}
}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
