% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.silhouette.ct.R
\name{plot.silhouette.ct}
\alias{plot.silhouette.ct}
\title{Plot a silhouette.ct object}
\usage{
\method{plot}{silhouette.ct}(x, mark.transitions = TRUE, xlab = "Time", ylab = "Silhouette", ...)
}
\arguments{
\item{x}{silhouette object produced by \code{\link{silhouette.ct}}}

\item{mark.transitions}{logical: Should transitions between clusters be marked
with vertical lines? Defaults to \code{TRUE}.}

\item{xlab, ylab}{x- and y-axis labels}

\item{\dots}{other arguments passed to \code{\link{plot}}}
}
\value{
None; a plot is generated.
}
\description{
Plots the silhouette index, generated by a call to
\code{\link{silhouette.ct}}, for a continuous-time k-means clustering object.
}
\seealso{
\code{\link{kmeans.ct}}, which includes an example; \code{\link{silhouette.ct}}
}
\author{
Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
