% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.ct.R
\name{standardize.ct}
\alias{standardize.ct}
\title{Center and scale a continuous-time multivariate data set}
\usage{
standardize.ct(fdobj)
}
\arguments{
\item{fdobj}{continuous-time multivariate data set of class \code{"\link[fda]{fd}"}}
}
\value{
A standardized (centered and scaled) version of the input data.
}
\description{
Subtracts the (continuous-time) mean and divides by the (continuous-time) standard
deviation of each of the variables. This is the continuous-time analogue
of taking an \eqn{n \times p} data matrix, subtracting the mean of each column, and 
dividing by the standard deviation of each column, as is done by 
\code{\link{scale}(\dots, center=TRUE, scale=TRUE)}.
}
\seealso{
\code{\link{center.ct}}
}
\author{
Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
