% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmva-package.R
\docType{package}
\name{ctmva-package}
\alias{ctmva-package}
\alias{ctmva}
\title{Continuous-Time Multivariate Analysis}
\description{
Implements continuous-time analogues of several classical techniques of multivariate analysis.
The inputs are \code{"\link[fda]{fd}"} (functional data) objects from the \pkg{fda} package.
}
\references{
Paul, Biplab; Reiss, Philip T.;  and Cui, Erjia (2023).
Continuous-time multivariate analysis. Preprint, doi.org/10.48550/arXiv.2307.09404.
}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
\keyword{package}
