% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmva-package.R
\docType{package}
\name{ctmva-package}
\alias{ctmva-package}
\alias{ctmva}
\title{Continuous-time multivariate analysis}
\description{
Implements continuous-time analogues of several classical techniques of multivariate analysis.
The inputs are \code{"\link[fda]{fd}"} (functional data) objects from the \pkg{fda} package.
}
\references{
Reiss, Philip T. and Paul, Biplab (2022). Continuous-time multivariate analysis. Preprint.

Ramsay, James O. and Silverman, Bernard W. (2005). Functional Data Analysis, 2nd ed., Springer, New York.

Ramsay, J. O.; Hooker, Giles and Graves, Spencer (2009). Functional Data Analysis with R and Matlab, Springer, New York.
}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Philip Tzvi Reiss <reiss@stat.haifa.ac.il>
}
\keyword{package}
